Test cleanup E: merge the three design tests into test_design.py
test_design_consume.py, test_design_fork_selector.py, and
test_design_jobserver.py each held a single small invariant test sharing the
same scaffolding. Combine them into one test_design.py with the three
TestCase classes intact and every scenario diagram preserved as a section
comment above its class. Same tests, same names, one file to navigate.

src/ line coverage unchanged.
